refactor(zcode): keep integration within size gates
