+++ src/shared/tui-agent-config-zcode.ts
@@ -1,0 +1,8 @@
+import type { TuiAgentConfig } from './tui-agent-config'
+
+export const ZCODE_TUI_AGENT_CONFIG: TuiAgentConfig = {
+  detectCmd: 'zcode',
+  launchCmd: 'zcode',
+  expectedProcess: 'zcode-cli',
+  promptInjectionMode: 'stdin-after-start'
+}
